auxilio

+++ src/main/java/co/edu/unbosque/catastromunicipal/domain/service/UrbanZoneService.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class UrbanZoneService {
@@ -15,16 +16,26 @@ public class UrbanZoneService {
     public List<UrbanZone> getAllUrbanZones(){
         return urbanZoneRepository.getAllUrbanZones();
     }
-    public List<UrbanZone> getUrbanZoneByName(String name){
+    public Optional<List<UrbanZone>> getUrbanZoneByName(String name){
         return urbanZoneRepository.getUrbanZoneByName(name);
     }
-    public void deleteUrbanZoneByName(String name){
-        urbanZoneRepository.deleteUrbanZoneByName(name);
+    public boolean deleteUrbanZoneByName(String name){
+        if(getUrbanZoneByName(name).isPresent()){
+            urbanZoneRepository.deleteUrbanZoneByName(name);
+            return true;
+        }else{
+            return false;
+        }
     }
     public  UrbanZone saveUrbanZone(UrbanZone urbanZone){
         return urbanZoneRepository.saveUrbanZone(urbanZone);
     }
-    public UrbanZone updateUrbanZone(UrbanZone urbanZone){
-        return urbanZoneRepository.updateUrbanZone(urbanZone);
+    public boolean updateUrbanZone(UrbanZone urbanZone){
+        if(getUrbanZoneByName(urbanZone.getZoneName()).isPresent()){
+            urbanZoneRepository.updateUrbanZone(urbanZone);
+            return true;
+        }else{
+            return false;
+        }
     }
 }
